Fixes inconsistencies in the Nrrd_Reader and Nrrd_Writer The Nrrd file handling is now compatible with ITKSnap and also 3D Slicer. Additionally the Conrad Grid settings, i.e. spacing and origin are set as expected and also the ImagePlus wrapping works fine. Thus, consistency is tested between ITKSnap, 3D Slicer, ImagePlus output (properties) and the Conrad Grid3D.
Bear in mind that the Nrrd files are no longer consistent with the official downloaded classes from Gregory Jefferis. The data will still load perfectly but the origin might be off as the sign is swapped. This was necessary as Gregory Jefferis classes have not been compatible with ITKSnap and also 3D Slicer, both being important software tools in the medical imaging community!